refactor: remove aditional .tsx in TriggersConfigs
