+++ frontend/src/components/configsList/TriggersConfigs.tsx
